Preliminary support for building on Watcom C 2.0.c http://sourceforge.net/projects/openwatcom/?source=typ_redirect
The Causeway DOS Extender supports DLL's so this is our 'way out'.  NULL builds, and loads up but crashes when we go to start the actual map.

I haven't debugged yet.

This fixes some things that the Watcom compiler was complaining about, and I verified I'm still able to build a working game with DJGPP