Move TransformVector x86 to r_misca.asm for Win32.
TODO:  Update VC6 and MinGW makefiles with this.  Convert to GAS.

Note: Goal of this branch is to move the inline win32 ASM to separate ASM files and finally get it into GAS.  There's quite a few additions that win32 has that non-win32 doesn't and it's bothered me for years.